moved backgrounds to HeaderFrame for toggle functions
